#10/component-wise implementation
