+++ client/src/components/leftmenu/leftmenu.tsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import { SettingsIcon, CloseIcon } from '@chakra-ui/icons';
+import { VStack, StackDivider, Box } from '@chakra-ui/react';
+import LeftMenuTab from './leftmenutab';
+import LeftMenuButton from './leftmenubutton';
+
+const LeftMenu = ({ menus }: any) => {
+  return (
+    <Box minH="100vh" maxW="150px" bg="white" ml="2">
+      <VStack
+        divider={<StackDivider borderColor="gray.200" />}
+        spacing={0}
+        align="stretch"
+      >
+        {menus.map((menu: any) => (
+          <LeftMenuTab
+            tabTitle={menu.tabTitle}
+            tabContents={menu.tabContents}
+          />
+        ))}
+      </VStack>
+      <Box>
+        <LeftMenuButton
+          name="Setting"
+          color="gray"
+          icon={<SettingsIcon w={3} h={3} />}
+        />
+        <LeftMenuButton
+          name="Quit"
+          color="red"
+          icon={<CloseIcon w={3} h={3} />}
+        />
+      </Box>
+    </Box>
+  );
+};
+
+export default LeftMenu;
